requirements.txt
boto3==1.26.69
botocore==1.29.69
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
flake8==6.0.0
Flask==2.2.2
idna==3.4
importlib-metadata==6.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
MarkupSafe==2.1.2
mccabe==0.7.0
pycodestyle==2.10.0
pyflakes==3.0.1
python-dateutil==2.8.2
python-dotenv==0.21.1
requests==2.28.2
s3transfer==0.6.0
six==1.16.0
urllib3==1.26.14
Werkzeug==2.2.2
zipp==3.12.1
